+++ packages/agents-a365-observability/src/tracing/scopes/OpenTelemetryScope.ts
@@ -88,7 +88,6 @@ export abstract class OpenTelemetryScope implements Disposable {
       this.setTagMaybe(OpenTelemetryConstants.GEN_AI_AGENT_NAME_KEY, agentDetails.agentName);
       this.setTagMaybe(OpenTelemetryConstants.GEN_AI_AGENT_DESCRIPTION_KEY, agentDetails.agentDescription);
       this.setTagMaybe(OpenTelemetryConstants.GEN_AI_AGENT_PLATFORM_ID_KEY, agentDetails.platformId);
-      this.setTagMaybe(OpenTelemetryConstants.GEN_AI_CONVERSATION_ID_KEY, agentDetails.conversationId);
       this.setTagMaybe(OpenTelemetryConstants.GEN_AI_ICON_URI_KEY, agentDetails.iconUri);
       this.setTagMaybe(OpenTelemetryConstants.GEN_AI_AGENT_AUID_KEY, agentDetails.agentAUID);
       this.setTagMaybe(OpenTelemetryConstants.GEN_AI_AGENT_UPN_KEY, agentDetails.agentUPN);
@@ -235,7 +234,7 @@ export abstract class OpenTelemetryScope implements Disposable {
       code: SpanStatusCode.ERROR,
       message
     });
-    this.span.setAttributes({ [OpenTelemetryConstants.ERROR_TYPE_KEY]: OpenTelemetryConstants.ERROR_TYPE_CANCELLED });
+    this.errorType = OpenTelemetryConstants.ERROR_TYPE_CANCELLED;
   }
 
   /**